Updates for front end
