+++ SkyFeedConnectAPI/src/main/resources/application.properties
@@ -11,10 +11,6 @@ spring.h2.console.path=/h2-console
 spring.h2.console.settings.trace=false
 spring.h2.console.settings.web-allow-others=false
 
-#spring.datasource.url=jdbc:postgresql://localhost:5432/project
-#spring.datasource.username=postgres
-#spring.datasource.password=admin
-
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
