+++ vite-project/src/assets/images/logo2.png
