+++ vite-project/src/components/App.jsx
@@ -42,11 +42,11 @@ const App = () => {
         <Route
           path="/main"
           element={
-            isAuthenticated ? (
-              <MainPage onLogout={handleLogout} />
-            ) : (
-              <Navigate replace to="/login" />
-            )
+            // isAuthenticated ? (
+            <MainPage onLogout={handleLogout} />
+            // ) : (
+            //   <Navigate replace to="/login" />
+            // )
           }
         />
       </Routes>
