+++ vite-project/src/components/features/NewsFeed/NewsFeed.jsx
@@ -1,13 +1,8 @@
-import "@style/NewsFeed.css";
-//import img1 from "../assets/alem-do-presidente.jpeg";
+import "./NewsFeed.css";
 import { useState, useEffect } from "react";
 import axios from "axios";
 
 const NewsFeed = () => {
-  // const[title, setTitle] = useState("")
-  // const[subtitle, setSubtitle] = useState("")
-  // const[summary, setSummary] = useState("")
-  // const[link, setLink] = useState("")
   const [newsItems, setNewsItems] = useState([]);
 
   const JWT_TOKEN = localStorage.getItem("userToken");
@@ -15,23 +10,18 @@ const NewsFeed = () => {
   useEffect(() => {
     const fetchNews = async () => {
       try {
-        const response = await axios.get(
-          `http://localhost:3333/newsIBGE/7`,
-
-          {
-            headers: {
-              Authorization: `Bearer ${JWT_TOKEN}`,
-            },
-          }
-        );
+        const response = await axios.get("http://localhost:3333/newsIBGE/7", {
+          headers: {
+            Authorization: `Bearer ${JWT_TOKEN}`,
+          },
+        });
         setNewsItems(response.data.newsList);
-        console.log(response.data.newsList);
       } catch (error) {
         console.error("Erro ao buscar notícias:", error);
       }
     };
     fetchNews();
-  });
+  }, []);
 
   return (
     <div className="news-feed">
@@ -39,6 +29,11 @@ const NewsFeed = () => {
         {newsItems.length > 0 && (
           <>
             <h1 className="main-news-subtitle">{newsItems[0].titulo}</h1>
+            <img
+              src={newsItems[0].imagens}
+              alt={newsItems[0].titulo}
+              className="secondary-news-image"
+            />
             <p className="main-news-summary">{newsItems[0].introducao}</p>
           </>
         )}
@@ -67,9 +62,9 @@ const NewsFeed = () => {
         {newsItems.slice(4).map((news) => (
           <div key={news.id} className="additional-news-item">
             <div className="additional-news-text">
-              <h2 className="secondary-news-title">{news.editorias}</h2>
-              <h3 className="secondary-news-subtitle">{news.titulo}</h3>
-              <p className="secondary-news-summary">{news.introducao}</p>
+              <h2 className="additional-news-title">{news.editorias}</h2>
+              <h3 className="additional-news-subtitle">{news.titulo}</h3>
+              <p className="additional-news-summary">{news.introducao}</p>
             </div>
           </div>
         ))}
