+++ vite-project/src/components/features/Weather/WeatherSearch.css
@@ -1,0 +1,98 @@
+#weather-content {
+  padding: 20px;
+  color: white;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+#weather-content h1.title {
+  font-size: 2.5rem;
+  margin-bottom: 3rem;
+  text-align: center;
+}
+
+.search-form {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  flex-direction: column;
+}
+
+.search-form input.inputWidget {
+  width: 100%;
+  padding: 10px;
+  border: 2px solid transparent;
+  border-radius: 4px;
+  color: #333;
+}
+
+#weather-button {
+  padding: 10px 15px;
+  background-color: #206be2;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+  width: 100%;
+  margin-top: 1rem;
+  margin-bottom: 4rem;
+}
+
+#weather-button:hover {
+  background-color: #59606b;
+}
+
+.tempo-data {
+  margin-top: 20px;
+  text-align: center;
+}
+
+.tempo-data h2 {
+  font-size: 2rem;
+  margin-bottom: 5px;
+}
+
+.tempo-data-info {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 10px;
+  flex-direction: column;
+}
+
+.tempo-data-info img {
+  width: 100px;
+  margin-right: 10px;
+}
+
+.tempo-data-info div {
+  text-align: center;
+  font-size: 2rem;
+  flex-direction: column;
+  display: flex;
+}
+
+.description::first-letter {
+  text-transform: uppercase;
+}
+
+.tempo-data-info .description {
+  margin-top: 20px;
+  margin-bottom: 10px;
+}
+
+.tempo-data-info .temperature {
+  margin-top: 20px;
+  margin-bottom: 10px;
+  font-size: 4rem;
+}
+
+.error-img {
+  max-width: 100%;
+  margin-top: 20px;
+  border-radius: 4px;
+}
