+++ vite-project/src/pages/LoginPage/LoginPage.css
@@ -1,22 +1,21 @@
-.auth-container {
+.sign-in-container {
   background: #fff;
+  max-width: 350px;
   padding: 40px;
   border-radius: 8px;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
   text-align: center;
   width: 100%;
-  max-width: 350px;
-  margin: auto;
   display: flex;
   flex-direction: column;
 }
 
-.auth-container h1 {
+.sign-in-container h1 {
   color: #333;
   margin-bottom: 1rem;
 }
 
-.subtitle {
+.sign-in-container p {
   color: #666;
   margin-bottom: 2rem;
 }
