+++ vite-project/src/pages/MainPage/MainPage.jsx
@@ -1,19 +1,20 @@
 import PropTypes from "prop-types";
-import "@style/MainPage.css";
-import WeatherWidget from "@components/WeatherSearch";
-import NewsFeed from "@components/NewsFeed";
-import logo from "../assets/logo2.png";
-import DateTimeWidget from "./DateTimeWidget";
+import { Link } from "react-router-dom";
+import "./MainPage.css";
+import WeatherWidget from "@features/Weather/WeatherSearch.jsx";
+import NewsFeed from "@features/NewsFeed/NewsFeed.jsx";
+import logo from "@images/logo2.png";
+import DateTimeWidget from "@features/DateTime/DateTimeWidget.jsx";
 import { RxExit } from "react-icons/rx";
 
 const MainPage = ({ onLogout }) => {
   const username = localStorage.getItem("userName") || "Usuário";
 
   return (
-    <section id="body">
-      <section id="header">
+    <div id="main-content" className="container">
+      <header className="header">
         <div className="header-content">
-          <div className="branch">
+          <div className="brand">
             <img src={logo} alt="Logo" className="logo" />
             <div className="header-title">
               <h1>SKY FEED CONNECT</h1>
@@ -22,23 +23,20 @@ const MainPage = ({ onLogout }) => {
           </div>
           <div className="header-user-info">
             <span className="username">{username}</span>
-            <a href="/login" onClick={onLogout} className="logout-link">
+            <Link to="/login" onClick={onLogout} className="logout-link">
               <RxExit />
-            </a>
+            </Link>
           </div>
         </div>
-      </section>
-      ,
-      <section id="principal">
-        <div className="personal-feed">
-          <NewsFeed className="news-feed" />
-          <div className="vertical-divider">
-            <WeatherWidget className="weather-widget" />
-            <DateTimeWidget />
-          </div>
-        </div>
-      </section>
-    </section>
+      </header>
+      <main id="principal">
+        <NewsFeed className="news-feed" />
+        <aside className="time-aside">
+          <WeatherWidget />
+          <DateTimeWidget />
+        </aside>
+      </main>
+    </div>
   );
 };
