--- vite-project/src/style/MainPage.css
@@ -1,93 +1,0 @@
-#header {
-  width: 100%;
-  margin-bottom: 30px;
-  display: flex;
-  justify-content: space-between;
-  background-color: #333333;
-  border-bottom: 1px solid #ccc;
-}
-
-.header-content {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: auto;
-  margin-top: 12px;
-  margin-bottom: 12px;
-  width: 1240px;
-  padding: 5px;
-  color: white;
-}
-
-.branch {
-  display: flex;
-  align-items: flex-end;
-  gap: 20px;
-}
-
-.header-title {
-  line-height: 0;
-}
-
-.header-title > h1 {
-  font-size: 36px;
-}
-
-.logo {
-  height: 90px;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.header-user-info {
-  display: flex;
-  align-items: center;
-}
-
-.username {
-  margin-right: 20px;
-  font-weight: bold;
-  color: white;
-}
-
-.logout-link {
-  color: white;
-  font-size: 30px;
-  cursor: pointer;
-  text-decoration: none;
-  align-items: center;
-  display: flex;
-  margin-right: 20px;
-}
-
-#principal .personal-feed {
-  background-color: white;
-  width: 1240px;
-  display: grid;
-  grid-template-columns: 70% 30%;
-  grid-template-areas: "A B";
-  justify-items: center;
-  margin: auto;
-}
-
-.vertical-divider {
-  background: linear-gradient(to bottom right, #69a3ff, #7b59ff);
-  padding: 20px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  height: 100%;
-  width: 100%;
-  font-family: "Arial", sans-serif;
-  color: #fff;
-  grid-area: B;
-}
-
-#page {
-  width: auto;
-  height: auto;
-}
-
-.news-feed {
-  grid-area: A;
-}
