--- vite-project/src/style/NewsFeed.css
@@ -1,85 +1,0 @@
-.main-news-item {
-  margin-bottom: 20px;
-}
-
-.main-news-item h1 {
-  width: 624px;
-  height: 201px;
-  font-weight: 700;
-  line-height: 60px;
-  color: #c4160c;
-  font-size: 52px;
-  padding-top: 3px;
-  margin-bottom: 20px;
-}
-
-.main-news-item p {
-  width: 625px;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 32px;
-  color: #303134;
-  padding-bottom: 7px;
-  border-bottom: 1px solid #ebedef;
-}
-
-.secondary-news-container {
-  width: 638.5px;
-  padding-top: 25px;
-  display: flex;
-  flex-direction: column;
-}
-
-.secondary-news-item {
-  width: 625px;
-  padding-bottom: 7px;
-  border-bottom: 1px solid #ebedef;
-  display: flex;
-  padding-left: 24px;
-}
-
-.secondary-news-container img {
-  width: 193px;
-  height: 110px;
-  border-radius: 12px;
-}
-
-.secondary-news-container h2 {
-  width: 192px;
-  height: 19px;
-  color: #c4170c;
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 17.5px;
-  padding-top: 11px;
-}
-
-.secondary-news-container h3 {
-  width: 194px;
-  height: 84px;
-  color: #2f3134;
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 25px;
-  margin-top: 16px;
-}
-
-.secondary-news-container p {
-  width: 174px;
-  height: 42.5px;
-
-  color: #2f3134;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 19px;
-}
-
-.additional-news-container {
-  display: flex;
-  margin-top: 25px;
-}
-
-.additional-news-item {
-  padding-left: 24px;
-  padding-bottom: 46px;
-}
