--- vite-project/src/style/WeatherSearch.css
@@ -1,89 +1,0 @@
-.weather-info {
-  display: flex;
-}
-
-#weather-content {
-  padding: 2rem;
-  background-color: #333;
-  border-radius: 10px;
-  color: white;
-}
-
-#weather-content .title {
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-}
-
-.weather-info-text {
-  flex: 1;
-}
-.weather-info-icon {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-}
-
-.weather-info-icon img {
-  width: 80px;
-}
-
-.weather-form {
-  display: flex;
-  width: 100%;
-}
-
-.weather-input {
-  flex: 1;
-  padding: 10px;
-  margin-right: 10px;
-  border: 2px solid #ddd;
-  border-radius: 8px;
-  background: #fff;
-  color: #000;
-}
-
-.weather-button {
-  padding: 10px;
-  border: none;
-  border-radius: 8px;
-  background-color: #007bff;
-  color: #fff;
-}
-
-.weather-button:hover {
-  background-color: #0056b3;
-}
-
-.error-img {
-  display: block;
-  max-width: 100%;
-  margin-top: 20px;
-  border-radius: 8px;
-}
-
-#datetime-widget {
-  background: linear-gradient(to bottom right, #69A3FF, #7B59FF);
-  border-radius: 12px;
-  padding: 20px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  width: 350px;
-  margin: 20px auto;
-  font-family: 'Arial', sans-serif;
-  color: #fff;
-  text-align: center;
-}
-
-.datetime-data h2, .datetime-data h3 {
-  margin: 0;
-}
-
-.tempo-data-info {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.tempo-data-info > span {
-  font-size: 3rem;
-  font-weight: bold;
-}
