--- vite-project/src/style/index.css
@@ -1,23 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-html,
-body {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-  align-items: center;
-  justify-content: center;
-}
-
-body {
-  font-family: -apple-system, BlinkMacSystemFont, "Encode Sans", sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  min-height: 100vh;
-}
-
-#root {
-  width: 100%;
-}
